Move derive interface
